requirements.txt
annotated-types==0.6.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.2
fuzzywuzzy==0.18.0
ibm-cloud-sdk-core==3.19.2
ibm-watson==8.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pdfminer.six==20231228
preshed==3.0.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.4.0
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
websocket-client==1.7.0
Werkzeug==3.0.1
